fix: 서버 파싱 인식 오류 해결

+++ Mallery4/app/src/main/java/com/example/mallery4/recyclerview/PostItem.kt
@@ -1,3 +1,3 @@
 package com.example.mallery4.recyclerview
 
-class PostItem(var groupName: String, var groupId: Long, var groupCount: String, var groupMembers: String, var groupNicknames: String, var postId: Int, var postImg: String, var postDate: String)
+class PostItem(var groupName: String, var groupId: Long, var groupCount: String, var groupMembers: String, var groupNicknames: String, var postId: Long, var postImg: String, var postDate: String)
